Prevent preprod emails being confused for prod emails
Add the option for ignoring specific email addresses - to be initially used to silence anything going to the enforcement team on preprod.

Add a large, bold message to the beginning of any email sent from a non-production system - identified by the presence of a domain whitelist - to say it's a test email.